+++ resources/views/artists/show.blade.php
@@ -11,23 +11,27 @@
 		<div class="row">
 			<div class="panel panel-default">
 				<div class="panel-heading collapse navbar-collapse">
-						<h2 class="col-md-4">{{ $artist->name }}</h2>
-						{!! Form::open([
-							'method' => 'DELETE',
-							'route' => ['artist.destroy', $artist->id],
-							'class'	=> 'navbar-form navbar-right prompt-confirm',
-							'msg'	=> trans('htmusic.are_you_sure')
-						]) !!}
-							{!! Form::submit(trans('htmusic.delete'), ['class' => 'btn btn-danger']) !!}
-						{!! Form::close() !!}
+					<h2 class="col-md-8">{{ $artist->name }}</h2>
+					@can('admin')
+						<div class="col-md-4">
+							{!! Form::open([
+								'method' => 'DELETE',
+								'route' => ['artist.destroy', $artist->id],
+								'class'	=> 'navbar-form navbar-right prompt-confirm',
+								'msg'	=> trans('htmusic.are_you_sure')
+							]) !!}
+								{!! Form::submit(trans('htmusic.delete'), ['class' => 'btn btn-danger']) !!}
+							{!! Form::close() !!}
 
-						{!! Form::open([
-							'method' => 'GET',
-							'route' => ['artist.edit', $artist->id],
-							'class'	=> 'navbar-form navbar-right',
-						]) !!}
-							{!! Form::submit(trans('htmusic.edit'), ['class' => 'btn btn-primary']) !!}
-						{!! Form::close() !!}
+							{!! Form::open([
+								'method' => 'GET',
+								'route' => ['artist.edit', $artist->id],
+								'class'	=> 'navbar-form navbar-right',
+							]) !!}
+								{!! Form::submit(trans('htmusic.edit'), ['class' => 'btn btn-primary']) !!}
+							{!! Form::close() !!}
+						</div>
+					@endcan
 				</div>
 				<div class="panel-body">
 					<div class="row">
